requirements.txt
python-dotenv
requests
psycopg2
simplejson
confluent_kafka
matplotlib
pandas
streamlit
kafka-python
streamlit-autorefresh